+++ src/ratalien/gameField.ts
@@ -275,6 +275,12 @@ export class GameField extends Control{
       const tile = this.map.toTileVector(point);//new Vector(Math.floor(point.x / 55), Math.floor(point.y / 55));
       this.objects.list.map(object => object.damage(point, tile, unit));
     }
+    unit.getResource = ()=>{
+        return this.objects.list.filter(it=> it.player!=unit.player && !(it instanceof Gold)) as InteractiveObject[];
+    }
+    unit.setTarget = (attackPoint) => {
+      this.setUnitTarget(unit, attackPoint);
+    }
     if (unit instanceof TruckUnit) {
       unit.getResource = ()=>{
         return this.objects.list.filter(it=> it instanceof Gold) as InteractiveObject[];
@@ -286,14 +292,7 @@ export class GameField extends Control{
       unit.getObjects = () => {
         return this.objects.list;
       }
-    } else if(unit instanceof SolderUnit) {
-       unit.getResource = ()=>{
-        return this.objects.list.filter(it=> it.player!=unit.player && !(it instanceof Gold)) as InteractiveObject[];
-      }
-       unit.setTarget = (attackPoint) => {
-        this.setUnitTarget(unit, attackPoint);
-      }
-    }
+    } 
     
     unit.player = player;
     //unit.position = new Vector(20, 20); //for demo
