+++ src/ratalien/units/DogUnit.ts
@@ -8,9 +8,18 @@ export class DogUnit extends AbstractUnit{
     this.weapon = new WeaponDog();
     this.weapon.onBulletTarget = (point)=>{
       this.onDamageTile?.(point);
+      const newEnemy = this.findClosestEnemy();
+      if (newEnemy !== this.targetEnemy) {
+        this.attackTarget = null;
+      }
     }
   }
 
+  step(delta: number) {
+    super.step(delta);
+    this.logic();
+  }
+  
   addGold(amount: number): boolean {
     return false;
   }
