+++ src/ratalien/units/HeavyTankUnit.ts
@@ -8,7 +8,16 @@ export class HeavyTankUnit extends AbstractUnit{
     this.weapon = new WeaponHeavyTank();
     this.weapon.onBulletTarget = (point)=>{
       this.onDamageTile?.(point);
+      const newEnemy = this.findClosestEnemy();
+      if (newEnemy !== this.targetEnemy) {
+        this.attackTarget = null;
+      }
     }
+  }
+
+  step(delta: number) {
+    super.step(delta);
+    this.logic();
   }
 
   addGold(amount: number): boolean {
