+++ src/ratalien/units/SolderUnit.ts
@@ -1,3 +1,4 @@
+import { Vector } from '../../common/vector';
 import { findClosestBuild, findClosestUnit } from '../distance';
 import {AbstractUnit} from './abstractUnit';
 import { InteractiveObject } from './interactiveObject';
@@ -5,7 +6,7 @@ import { MapObject } from './mapObject';
 import { WeaponSolder } from './weaponSolder';
 export class SolderUnit extends AbstractUnit{
   weapon: WeaponSolder;
-  targetEnemy: { distance: number; unit: AbstractUnit; } | { distance: number; unit: MapObject; tile: import("c:/Users/админ/Desktop/rs/rs-clone2/artquiz/src/common/vector").Vector; };
+ 
   constructor(){
     super();
     this.weapon = new WeaponSolder();
@@ -30,28 +31,7 @@ export class SolderUnit extends AbstractUnit{
     return 10;    
   }
 
-  logic() {
-    const closestUnit = findClosestUnit(this.position.clone(), this.getResource().filter(it => it instanceof AbstractUnit) as AbstractUnit[]);
-    const closestBuild = findClosestBuild(this.position.clone(), this.getResource().filter(it => it instanceof MapObject) as MapObject[]);
-    const targetEnemy = closestUnit.distance > closestBuild.distance ? closestBuild : closestUnit;
-   
-    if (!this.attackTarget) {
-      this.targetEnemy = targetEnemy;
-      if (this.targetEnemy.unit instanceof MapObject) {
-        this.setTarget(closestBuild.tile)
-      } else {
-        this.setTarget(closestUnit.unit.position);
-      }
-    } else if(this.targetEnemy.unit.health ===0) {
-      this.attackTarget = null;
-    }
-  }
-
-  findClosestEnemy() {
-    const closestUnit = findClosestUnit(this.position.clone(), this.getResource().filter(it => it instanceof AbstractUnit) as AbstractUnit[]);
-    const closestBuild = findClosestBuild(this.position.clone(), this.getResource().filter(it => it instanceof MapObject) as MapObject[]);
-    return closestUnit.distance > closestBuild.distance ? closestBuild : closestUnit;
-  }
+  
 
 
 }
