+++ src/ratalien/units/TankUnit.ts
@@ -8,7 +8,16 @@ export class TankUnit extends AbstractUnit{
     this.weapon = new WeaponTank();
     this.weapon.onBulletTarget = (point)=>{
       this.onDamageTile?.(point);
+      const newEnemy = this.findClosestEnemy();
+      if (newEnemy !== this.targetEnemy) {
+        this.attackTarget = null;
+      }
     }
+  }
+
+  step(delta: number) {
+    super.step(delta);
+    this.logic();
   }
 
   addGold(amount: number): boolean {
