+++ src/ratalien/units/abstractUnit.ts
@@ -1,7 +1,9 @@
-import {Vector, IVector} from "../../common/vector";
+import { findClosestBuild, findClosestUnit } from '../distance';
+import { Vector, IVector } from "../../common/vector";
 import { InteractiveObject } from "./interactiveObject";
 import {consts} from "../globals";
 import AbstractWeapon from "./abstractWeapon";
+import { MapObject } from './mapObject';
 
 export class AbstractUnit extends InteractiveObject{
   positionPx: Vector;
@@ -23,6 +25,7 @@ export class AbstractUnit extends InteractiveObject{
   protected gold: number = 0;
   maxGold: number = 3000;
   weapon: AbstractWeapon;
+   targetEnemy: { distance: number; unit: AbstractUnit; } | { distance: number; unit: MapObject; tile: Vector; };
 
   get position(){
     return new Vector(Math.floor(this.positionPx.x/55), Math.floor(this.positionPx.y / 55));
@@ -153,8 +156,27 @@ export class AbstractUnit extends InteractiveObject{
     return action;
   }
 
-  logic() {
-    
+ logic() {
+    const closestUnit = findClosestUnit(this.position.clone(), this.getResource().filter(it => it instanceof AbstractUnit) as AbstractUnit[]);
+    const closestBuild = findClosestBuild(this.position.clone(), this.getResource().filter(it => it instanceof MapObject) as MapObject[]);
+    const targetEnemy = closestUnit.distance > closestBuild.distance ? closestBuild : closestUnit;
+   
+    if (!this.attackTarget) {
+      this.targetEnemy = targetEnemy;
+      if (this.targetEnemy.unit instanceof MapObject) {
+        this.setTarget(closestBuild.tile)
+      } else {
+        this.setTarget(closestUnit.unit.position);
+      }
+    } else if(this.targetEnemy.unit.health ===0) {
+      this.attackTarget = null;
+    }
+  }
+
+  findClosestEnemy() {
+    const closestUnit = findClosestUnit(this.position.clone(), this.getResource().filter(it => it instanceof AbstractUnit) as AbstractUnit[]);
+    const closestBuild = findClosestBuild(this.position.clone(), this.getResource().filter(it => it instanceof MapObject) as MapObject[]);
+    return closestUnit.distance > closestBuild.distance ? closestBuild : closestUnit;
   }
